Move BrowserCapabilities to stripe-core.
