+++ stripe-core/build.gradle
@@ -27,6 +27,7 @@ dependencies {
     kapt "com.google.dagger:dagger-compiler:$daggerVersion"
 
     implementation "androidx.activity:activity-compose:$androidxActivityVersion"
+    implementation "androidx.browser:browser:$androidxBrowserVersion"
 
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$kotlinCoroutinesVersion"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$kotlinCoroutinesVersion"
