+++ stripe-core/src/main/java/com/stripe/android/core/browser/BrowserCapabilities.kt
@@ -1,10 +1,13 @@
-package com.stripe.android.payments
+package com.stripe.android.core.browser
+
+import androidx.annotation.RestrictTo
 
 /**
  * Representation of the device's browser capabilities. Used for determining how to handle
  * browser-based payment authentication.
  */
-internal enum class BrowserCapabilities {
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+enum class BrowserCapabilities {
     CustomTabs,
     Unknown
 }
